requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.1

google-genai>=1.0.0
weaviate-client>=4.0.0
